requirements.txt
django
django-rest-swagger
djangorestframework-jwt
drf-yasg
